+++ src/control/ScrollHandler.h
@@ -35,16 +35,16 @@ public:
 	void goToFirstPage();
 
 	void scrollToPage(PageRef page, double top = 0);
-	void scrollToPage(int page, double top = 0);
+	void scrollToPage(size_t page, double top = 0);
 
 	void scrollToSpinPange();
 
 	void scrollToAnnotatedPage(bool next);
 
-	bool isPageVisible(int page, int* visibleHeight = NULL);
+	bool isPageVisible(size_t page, int* visibleHeight = NULL);
 
 public:
-	virtual void pageChanged(int page);
+	virtual void pageChanged(size_t page);
 
 private:
 	XOJ_TYPE_ATTRIB;
